+++ CHANGELOG.md
@@ -6,8 +6,16 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [2.2.0] - 2017-08-11
+### [Added]
+- Change option for default cluster mode enabled
+- Recreate session when error in thread keepalive
+- Add method for test if Janus response it's a error
+- Test Janus instance is down and recreate when instance is start
+
 ### [Fixed]
+- Create session only cluster mode is disabled
 - Cleaning cluster classes
+- Get settings to cluster mode
 
 ## [2.1.1] - 2017-06-14
 ### [Fixed]